Move transaction filter into models
